Add files via upload
This is my grade 12 Computer Science Culminating Project (Using Java), here I have made a simple Pokémon battle game that is run within the Terminal. It incorporates various aspects of Java functionality including classes, methods, loops, objects and more. 

The PokemonTester.java file is used to run the overall game, PokemonPlayer.java is the class which is for the player object used by the user and the computer, Finally Pokemon.java is the class which is used for each individual Pokémon object.